Implemented user-fetching-optimization and took the caching and handling getting users logic out of the component
