+++ src/App.tsx
@@ -1,79 +1,45 @@
-import { FC, useEffect, useState } from "react";
+import { FC, useState } from "react";
 import "./styles/index.scss";
 import Container from "./components/Container/Container";
 import TableHead from "./components/TableHead/TableHead";
 import Pagination from "./components/Pagination/Pagintaion";
 import { IUser } from "./interface/user.interface";
-import UserService from "./services/UserService";
 import Controls from "./components/Controls/Controls";
 import UsersList from "./components/Users/UsersList";
 import Error from "./components/Error/Error";
 import Modal from "./components/Modal/Modal";
-
-interface SavedData {
-  [currentPage: number]: IUser[];
-}
+import { useGetUsers } from "./hooks/useGetUsers";
+import { useToggle } from "./hooks/useToggle";
+import { useModal } from "./hooks/useModal";
 
 const App: FC = () => {
-  const [users, setUsers] = useState<IUser[]>([]);
-  const [cachedUsers, setCachedUsers] = useState<SavedData>({});
-  const [error, setError] = useState<string>("");
-  const [loading, setLoading] = useState<boolean>(false);
-  const [isOpenModal, setIsOpenModal] = useState<boolean>(false);
   const [chosenUser, setChosenUser] = useState<IUser>(null);
-
   const [currentPage, setCurrentPage] = useState<number>(1);
   const [nat, setNat] = useState<string>("");
-  const [clickFilter, setClickFilter] = useState<boolean>(false);
   const limit = 10;
   const totalUsers = 10000;
 
-  const getUsers = async () => {
-    try {
-      setLoading(true);
-
-      const response = await UserService.getUsers({
-        nat,
-        limit,
-        page: currentPage,
-      });
-
-      const data = response.data.results;
-      setUsers(data);
-
-      if (nat === "")
-        setCachedUsers((prev) => ({ ...prev, [currentPage]: data }));
-    } catch (err) {
-      setError(err.message);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    const hasCachedUsers = cachedUsers[currentPage]?.length > 0;
-    if (nat === "" && hasCachedUsers) {
-      setUsers(cachedUsers[currentPage]);
-    }
-    if (nat !== "" || (nat === "" && !hasCachedUsers)) {
-      getUsers();
-    }
-  }, [currentPage, clickFilter]);
+  const filter = useToggle(false);
+  const modal = useModal();
+  const { users, error, loading } = useGetUsers(
+    nat,
+    filter.toggle,
+    currentPage,
+    limit
+  );
 
   const onClickDropdownItem = (item: string) =>
     item === "Clear" ? setNat("") : setNat(item);
 
   const onClickPage = (num: number) => setCurrentPage(num);
 
-  const onClickFilter = () => setClickFilter(!clickFilter);
+  const onClickFilter = () => filter.onClickToggle();
 
   const onClickUser = (idValue: number | string) => {
     const user = users.find((user) => user.id.value === idValue);
     setChosenUser(user);
-    setIsOpenModal(true);
+    modal.toggleModal();
   };
-
-  const onCloseModal = () => setIsOpenModal(false);
 
   return (
     <div className="app">
@@ -95,7 +61,9 @@ const App: FC = () => {
           currentPage={currentPage}
           onClickPage={onClickPage}
         />
-        {isOpenModal && <Modal onCloseModal={onCloseModal} user={chosenUser} />}
+        {modal.modalOpen && (
+          <Modal onCloseModal={modal.toggleModal} user={chosenUser} />
+        )}
       </Container>
     </div>
   );
