+++ src/hooks/useModal.ts
@@ -1,0 +1,9 @@
+import { useState } from "react";
+
+export function useModal() {
+  const [modalOpen, setModalOpen] = useState<boolean>(false);
+
+  const toggleModal = () => setModalOpen(!modalOpen);
+
+  return { modalOpen, toggleModal }
+}
