+++ src/interface/savedUsers.interface.ts
@@ -1,0 +1,5 @@
+import { IUser } from "./user.interface";
+
+export interface SavedData {
+  [currentPage: number]: IUser[];
+}
